Switched the order of i18n:attributes values with explicit message ids.
Previously, it was expected that when providing both the attribute name and
the messsage id the first value be the message id and the second one the name
of the attribute.

However, at least the Chameleon implementation of ZPT implements this order
differently (and IIRC the zope.pagetemplate does the same) requiring the
first item to be the name of the attribute and the second (optional) one
the message id.

See http://pagetemplates.org/docs/latest/reference.html#i18n-attributes
for details.